bpgv_strategy/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
fredapi>=0.5.0
arch>=6.2.0
yfinance>=0.2.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
python-dotenv>=1.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
scipy>=1.10.0